Crate synthax

Synthesize syntax with quasiquoting plugins.

The quasiquoting plugins supplied by this crate behave like those supplied by libsyntax (i.e., #![feature(quote)]) with two differences.

First, the results of expressions can be quoted directly without intermediate variables by enclosing the expression in ${..}. For example, to quote the first element of the slice tokens, write ${tokens[0]}.

let tokens = &[Token::Eq, Token::Semi];
let stmt = quote_stmt!(context, let foo ${tokens[0]} bar ${tokens[1]});Run

Second, slices and Vecs containing quotable values can be quoted in sequences. The sequences are written like and behave like the sequences in macros. For example, to quote the values of the slice idents with , between each quoted value, write $($ident), *.

let idents = &[context.ident_of("foo"), context.ident_of("bar"), context.ident_of("baz")];
let pat = quote_pat!(context, Struct { $(ref $idents), * });
let pat = quote_pat!(context, Struct { $(ref mut ${&idents[..1]}), * });Run

Plugins

quote_token_trees!

quote_arg!

quote_arm!

quote_attribute!

quote_bare_fn_ty!

quote_block!

quote_expr!

quote_field!

quote_field_pat!

quote_fn_decl!

quote_foreign_item!

quote_foreign_mod!

quote_generics!

quote_impl_item!

quote_item!

quote_lifetime!

quote_lifetime_defs!

quote_lit!

quote_local!

quote_meta_item!

quote_pat!

quote_path!

quote_stmt!

quote_struct_field!

quote_trait_item!

quote_ty!

quote_variant!

quote_where_clause!

Structs

Sequence
A quoted sequence.

SequenceBuilder
Constructs sequences.

Traits

SequenceQuote
A type quotable in a sequence.

ToExpr
An AST type whose values can be generated by evaluating expressions.

ToTokenTrees
An AST type whose values can be converted into a stream of token trees.