Trait synthax::ToExpr

pub trait ToExpr {
    fn to_expr(&self, context: &ExtCtxt, span: Span) -> P<Expr>;
}

An AST type whose values can be generated by evaluating expressions.

Required Methods

fn to_expr(&self, context: &ExtCtxt, span: Span) -> P<Expr>

Returns an expression which would produce this value if evaluated.